uib-var component moved out of unnecessary sub-folder
