+++ src/front-end/readme.md
@@ -1,0 +1,6 @@
+This is the source code for the old `uibuilderfe` front end
+library.
+
+It is "functionally stabilised" and no longer receiving updates.
+It will eventually be deprecated and then removed in a future
+major release.
